chore: reorganize tests and split tsconfigs
Made-with: Cursor
