+++ package.json
@@ -18,7 +18,7 @@
     "test:manual:watch": "run-p dev test:manual:wait",
     "clean": "rimraf build",
     "check:types": "run-p check:types:src check:types:test",
-    "check:types:src": "tsc",
+    "check:types:src": "tsc -p tsconfig.src.json",
     "check:types:test": "tsc -p tsconfig.test.json",
     "reset": "rimraf node_modules",
     "test": "jest",
