+++ src/background/requestHandlers/__tests__/moveBookmarkNodeHandler.test.ts
@@ -6,7 +6,7 @@ jest.mock("@/shared/lib/browser/bookmarks", () => ({
   },
 }));
 
-import moveBookmarkNodeHandler from "./moveBookmarkNodeHandler";
+import moveBookmarkNodeHandler from "../moveBookmarkNodeHandler";
 
 function flushAsync() {
   return new Promise((resolve) => setImmediate(resolve));
